Corrección del peso en las mascotas

+++ backend/patitas-care/src/main/java/com/kmd/patitas_care/domain/model/entity/Mascota.java
@@ -24,7 +24,7 @@ public class Mascota {
     private String especie;
     private String raza;
     private int edad;
-    private int peso;
+    private double peso;
     private int size;
     private String color;
 }
